Address compiler warnings raised with standard and stricter options:     - addressed all warnings produced by standard use of Clang and MSVS     - fixed many of narrowing warnings not reported by default with GCC     - replaced many non-unit bitfields with small int types and casts to       reduce GCC warnings with -Wconversion

+++ regression/bfr_evaluate/bfrSurfaceEvaluator.cpp
@@ -52,7 +52,7 @@ BfrSurfaceEvaluator<REAL>::Evaluate(Bfr::Index                baseFace,
                                     EvalResults<REAL>       & results) const {
 
     //  Allocate vectors for the properties to be evaluated:
-    int numCoords = tessCoords.size() / 2;
+    int numCoords = (int) tessCoords.size() / 2;
 
     results.Resize(numCoords);
